requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
confluent-kafka==2.13.2
elastic-transport==8.13.0
elasticsearch==8.13.0
idna==3.11
kafka-python-ng==2.2.3
mysql-connector-python==9.6.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3